# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi MATKUL_API (UNESA)",
        "description": "Dokumentasi API untuk proyek MATKUL_API. Mencakup Dosen, Mahasiswa, dll. Dilindungi oleh JWT.",
        "contact": {
            "email": "admin@unesa.ac.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Lokal (http://127.0.0.1:8000)"
        }
    ],
    "paths": {
...
        "/api/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil daftar semua mahasiswa (dengan opsi filter nama)",
                "description": "Mengembalikan daftar semua data mahasiswa. Dapat difilter berdasarkan nama. Membutuhkan autentikasi JWT.",
                "operationId": "getMahasiswaList",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Filter berdasarkan nama mahasiswa (pencarian sebagian, case-insensitive)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mahasiswa"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Buat data mahasiswa baru",
                "description": "Menyimpan data mahasiswa baru ke dalam database. Membutuhkan autentikasi JWT.",
                "operationId": "storeMahasiswa",
                "requestBody": {
                    "description": "Payload data mahasiswa untuk dibuat",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MahasiswaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mahasiswa berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input tidak valid / Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mahasiswa/{nim}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil data mahasiswa berdasarkan NIM",
                "description": "Mengembalikan satu data mahasiswa berdasarkan NIM. Membutuhkan autentikasi JWT.",
                "operationId": "getMahasiswaByNim",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token tidak valid"
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Update data mahasiswa berdasarkan NIM",
                "description": "Memperbarui data mahasiswa berdasarkan NIM. Field yang tidak disertakan atau null di payload tidak akan diubah (karena validasi `sometimes`). Membutuhkan autentikasi JWT.",
                "operationId": "updateMahasiswaByNim",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payload data mahasiswa untuk update (field opsional)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MahasiswaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mahasiswa berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mahasiswa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input tidak valid / Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token tidak valid"
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus data mahasiswa berdasarkan NIM",
                "description": "Menghapus data mahasiswa berdasarkan NIM. Membutuhkan autentikasi JWT.",
                "operationId": "deleteMahasiswaByNim",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mahasiswa berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mahasiswa berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token tidak valid"
                    },
                    "404": {
                        "description": "Mahasiswa tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Mahasiswa": {
                "title": "Mahasiswa Model",
                "description": "Representasi data mahasiswa",
                "required": [
                    "id",
                    "nim",
                    "nama"
                ],
                "properties": {
                    "id": {
                        "description": "ID unik internal",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nim": {
                        "description": "Nomor Induk Mahasiswa (digunakan sebagai ID di route)",
                        "type": "string",
                        "example": "18051204001"
                    },
                    "nama": {
                        "description": "Nama lengkap mahasiswa",
                        "type": "string",
                        "example": "Adi Santoso"
                    },
                    "jenis_kelamin": {
                        "description": "Jenis kelamin (L/P)",
                        "type": "string",
                        "enum": [
                            "L",
                            "P"
                        ],
                        "example": "L"
                    },
                    "alamat": {
                        "description": "Alamat tinggal mahasiswa",
                        "type": "string",
                        "example": "Jl. Pendidikan No. 5"
                    },
                    "tanggal_lahir": {
                        "description": "Tanggal lahir mahasiswa (YYYY-MM-DD)",
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "program_studi": {
                        "description": "Program studi",
                        "type": "string",
                        "example": "S1 Teknik Informatika"
                    },
                    "angkatan": {
                        "description": "Tahun angkatan",
                        "type": "integer",
                        "example": 2018
                    },
                    "email": {
                        "description": "Alamat email mahasiswa",
                        "type": "string",
                        "format": "email",
                        "example": "adi.santoso@example.com"
                    },
                    "created_at": {
                        "description": "Waktu data dibuat",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Waktu data terakhir diupdate",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MahasiswaInput": {
                "title": "Mahasiswa Input Payload",
                "description": "Data yang dibutuhkan untuk membuat atau mengupdate mahasiswa",
                "required": [
                    "nim",
                    "nama",
                    "jenis_kelamin",
                    "alamat",
                    "tanggal_lahir",
                    "program_studi",
                    "angkatan",
                    "email"
                ],
                "properties": {
                    "nim": {
                        "type": "string",
                        "example": "19051204002"
                    },
                    "nama": {
                        "type": "string",
                        "example": "Budi Raharjo"
                    },
                    "jenis_kelamin": {
                        "type": "string",
                        "enum": [
                            "L",
                            "P"
                        ],
                        "example": "P"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Teknologi No. 12"
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "2001-05-20"
                    },
                    "program_studi": {
                        "type": "string",
                        "example": "S1 Sistem Informasi"
                    },
                    "angkatan": {
                        "type": "integer",
                        "example": 2019
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "budi.raharjo@example.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dengan prefix Bearer (Contoh: 'Bearer token_anda_disini')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}